Removed some of the dummy code. Fixed few minor bugs caused by typos.
As we are now moving to utilize real data from mongoDB, I removed previously used dummy objects and made needed changes in render statements to correctly point to the results fetched from mongoDB. There were some wrong variable names which were corrected. Also attempted to find a way to decode the URLs (source and image) but so far it has been unsuccesfull. Also tags are not saved correctly and I was unable to fix this. Additionally in comic details view it does seem like showing tags doesn't work as intended as the template attempts to show tags even if the tag array is empty.