Epic fight with bootstrap grid layout.
Hammering the bootstrap grid to flip the comic sorting form on top of the listing on comic listing page on small screens almost works. It just breaks with a certain screen size. Whatever. Not into fixing it any further. Also added comments etc. acroos the other files included in the commit.